requirements.txt
pyyaml
pymongo
pymysql
pyfiglet
